chore: #147 폴더정리 1차
