+++ ByeBoo-iOS/ByeBoo-iOS/Presentation/Feature/Home/View/Onboarding/JourneyResultView.swift
@@ -7,42 +7,6 @@
 
 import UIKit
 
-enum JourneyType: String {
-    case face = "감정 직면"
-    case process = "감정 정리"
-    
-    var image: UIImage {
-        switch self {
-        case .face:
-                .faceEmotion
-        case .process:
-                .processEmotion
-        }
-    }
-    
-    var description: String {
-        return "\(rawValue) 여정"
-    }
-    
-    var frontImage: UIImage {
-        switch self {
-        case .face:
-                .faceFrontCard
-        case .process:
-                .processFrontCard
-        }
-    }
-    
-    var backImage: UIImage {
-        switch self {
-        case .face:
-                .faceBackCard
-        case .process:
-                .processBackCard
-        }
-    }
-}
-
 final class JourneyResultView: BaseView {
 
     private let backgroundImageView = UIImageView()
