+++ ByeBoo-iOS/ByeBoo-iOS/Presentation/Feature/Information/Common/InformationBaseView.swift
