+++ ByeBoo-iOS/ByeBoo-iOS/Presentation/Feature/Information/Common/InformationViewType.swift
