+++ ByeBoo-iOS/ByeBoo-iOS/Presentation/Feature/Quest/ViewModel/QuestsViewModel.swift
@@ -7,7 +7,7 @@
 
 import Combine
 
-final class QuestsViewModel: ViewModelType {
+final class QuestsViewModel {
     
     private let cancellables = Set<AnyCancellable>()
     private let nameSubject = PassthroughSubject<Result<String, ByeBooError>, Never>.init()
@@ -84,9 +84,9 @@ final class QuestsViewModel: ViewModelType {
     }
 }
 
-extension QuestsViewModel {
+extension QuestsViewModel: ViewModelType {
     
-    enum InputAction {
+    enum Input {
         case questViewWillAppear
         case questLoadView
     }
@@ -98,7 +98,7 @@ extension QuestsViewModel {
         let loadingPublisher: AnyPublisher<Bool, Never>
     }
     
-    func action(_ trigger: InputAction) {
+    func action(_ trigger: Input) {
         switch trigger {
         case .questViewWillAppear:
             getUseName()
